requirements.txt
typing
pytest
pywin32
click
python-docx